+++ administrator/sql/install.mysql.utf8.sql
@@ -12,10 +12,12 @@ CREATE TABLE IF NOT EXISTS `#__web357test_recipes` (
 `ingredients` VARCHAR(255)  NOT NULL ,
 `cooking_time` DOUBLE NULL ,
 `difficulty` VARCHAR(255)  NULL  DEFAULT "",
+`serving_size` TINYINT UNSIGNED NOT NULL DEFAULT 10,
 PRIMARY KEY (`id`)
 ,KEY `idx_state` (`state`)
 ,KEY `idx_checked_out` (`checked_out`)
 ,KEY `idx_created_by` (`created_by`)
 ,KEY `idx_modified_by` (`modified_by`)
+,KEY `idx_serving_size` (`serving_size`)
 ) DEFAULT COLLATE=utf8mb4_unicode_ci;