+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
+## [Unreleased]
+
+### Added
+
+- New parameter "Serving size" for Recipe in the administrator form